requirements.txt
# Core dependencies for snowflower content engine
google-api-python-client>=2.140
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2
requests>=2.32
atproto>=0.0.50
python-dotenv>=1.0
typer>=0.12
pydantic>=2.7
rich>=13.7
pyyaml>=6.0

# Dev / testing
pytest>=8.2
pytest-cov>=5.0
ruff>=0.6

# Video / captions / images (optional — install as needed)
# ffmpeg-python>=0.2
# faster-whisper>=1.0
# Pillow>=10.4

# GenAI (optional)
# elevenlabs>=1.5
# fal-client>=0.5
